CIF-1783 - Implement ComponentExporter for product carousel component
Write a unit test for the `getComponentData` method.
